Implement base entity
