+++ homeassistant/components/yale_smart_alarm/binary_sensor.py
@@ -6,14 +6,12 @@ from homeassistant.components.binary_sensor import (
     BinarySensorEntity,
 )
 from homeassistant.config_entries import ConfigEntry
-from homeassistant.const import CONF_USERNAME
 from homeassistant.core import HomeAssistant
-from homeassistant.helpers.entity import DeviceInfo
 from homeassistant.helpers.entity_platform import AddEntitiesCallback
-from homeassistant.helpers.update_coordinator import CoordinatorEntity
 
-from .const import COORDINATOR, DOMAIN, MANUFACTURER, MODEL
+from .const import COORDINATOR, DOMAIN
 from .coordinator import YaleDataUpdateCoordinator
+from .entity import YaleEntity
 
 
 async def async_setup_entry(
@@ -30,24 +28,10 @@ async def async_setup_entry(
     )
 
 
-class YaleBinarySensor(CoordinatorEntity, BinarySensorEntity):
+class YaleBinarySensor(YaleEntity, BinarySensorEntity):
     """Representation of a Yale binary sensor."""
 
     _attr_device_class = BinarySensorDeviceClass.DOOR
-
-    def __init__(self, coordinator: YaleDataUpdateCoordinator, data: dict) -> None:
-        """Initialize the Yale Lock Device."""
-        super().__init__(coordinator)
-        self._coordinator = coordinator
-        self._attr_name = data["name"]
-        self._attr_unique_id = data["address"]
-        self._attr_device_info = DeviceInfo(
-            name=self._attr_name,
-            manufacturer=MANUFACTURER,
-            model=MODEL,
-            identifiers={(DOMAIN, data["address"])},
-            via_device=(DOMAIN, self._coordinator.entry.data[CONF_USERNAME]),
-        )
 
     @property
     def is_on(self) -> bool | None:
